rewrite index.html using the new page layout
The index page was hardly a representative of our needs. Half the page
talked about Sinatra, and the other half containing the navigation links
was rendered redundant due to the new page layout introduced in SHA:
b4431589. Removed the navigation links, and replaced the remaining cruft
with a small description of KLUG.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>